pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "visionlab-datasets"         # <-- avoid conflict with HF 'datasets'
version = "0.4.0"
description = "Visionlab dataset registry and preparation utilities for slipstream"
readme = "README.md"
authors = [
    {name = "George Alvarez", email = "grez72@gmail.com"}
]
requires-python = ">=3.10,<3.13"

dependencies = [
    # Slipstream data loader
    "visionlab-slipstream@git+https://github.com/harvard-visionlab/slipstream.git@main",

    # Torch stack (for preprocessing transforms)
    "torch>=2.1.0",
    "torchvision>=0.16.0",

    # Image utilities
    "pillow",
    "numpy",
]

[project.optional-dependencies]
# Legacy: for projects still using StreamingDatasetVisionlab
legacy = [
    "litdata@git+https://github.com/Lightning-AI/litdata.git@main",
    "remote-data@git+https://github.com/harvard-visionlab/remote-data.git@main",
    "visionlab-auth@git+https://github.com/harvard-visionlab/auth.git@main",
    "tqdm",
]

[dependency-groups]
dev = [
    "pytest",
    "pytest-xdist",
    "ruff",
    "ipykernel",
    "nbstripout",
    "ipywidgets",
    "pandas",
    "seaborn",
    "matplotlib",
    "jupyterlab",
]

[tool.setuptools]
# We map the top-level import 'visionlab.datasets' to your folder 'datasets'
package-dir = { "visionlab.datasets" = "datasets" }
packages = [
  "visionlab.datasets",
  "visionlab.datasets._configs",
  "visionlab.datasets.prep",
]
